backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
package-mode = false
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "alembic (>=1.15.2,<2.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "boto3 (>=1.38.3,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi (>=0.104.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "pytest-anyio (>=0.0.0,<0.0.1)",
    "anyio (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "fastapi-limiter (>=0.1.6)",
    "redis (>=5.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pytest = "^8.3.5"
httpx = "^0.28.1"
pillow = "^11.2.1"
pytest-mock = "^3.14.0"

